requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
python-nmap==0.7.1
requests==2.32.3
urllib3==2.4.0
